+++ application/controllers/Komponen.php
@@ -48,7 +48,7 @@ class Komponen extends CI_Controller {
 	 * Load Page Insert
 	 * @return [type] [description]
 	 */
-	public function insert()
+	public function insert($id_perangkat)
 	{
 		//set navbar
 		$this->template->set_navbar('templates/navbar');
@@ -56,7 +56,7 @@ class Komponen extends CI_Controller {
 		//set rules form validation
 		$this->_set_validation();
 		//get perangkat list
-		$data['perangkat'] = $this->perangkat_model->get_all();
+		$data['perangkat'] = $this->perangkat_model->get_one(array('id'=> $id_perangkat));
 
 		if($this->form_validation->run() == FALSE)
 		{
@@ -80,7 +80,7 @@ class Komponen extends CI_Controller {
 				$this->session->set_flashdata('message_danger', "Failed inserting data");	
 			}
 
-			redirect('komponen');
+			redirect('perangkat/view/'.$id_perangkat);
 			
 		}
 	}
@@ -89,9 +89,9 @@ class Komponen extends CI_Controller {
 	 * Load Page Update
 	 * @return [type] [description]
 	 */
-	public function update($id)
+	public function update($id_perangkat, $id)
 	{
-		if($id) {
+		if($id && $id_perangkat) {
 			//load helper validation to set form
 			$this->load->helper('validation');
 
@@ -101,7 +101,7 @@ class Komponen extends CI_Controller {
 			//get a record by id
 			$data['model'] = $this->komponen_model->get_one($condition);
 			//get perangkat list
-			$data['perangkat'] = $this->perangkat_model->get_all();
+			$data['perangkat'] = $this->perangkat_model->get_one(array('id'=>$id_perangkat));
 
 			//set navbar in template
 			$this->template->set_navbar('templates/navbar');
@@ -133,12 +133,12 @@ class Komponen extends CI_Controller {
 					$this->session->set_flashdata('message_danger', "Failed updating data");	
 				}
 
-				redirect('komponen');
+				redirect('perangkat/view/'.$id_perangkat);
 				
 			}	
 		} else {
 			$this->session->set_flashdata('message_danger', "System Error");
-			redirect('komponen');
+			redirect('perangkat');
 		}
 	}
 
@@ -146,9 +146,9 @@ class Komponen extends CI_Controller {
 	 * Load Page View
 	 * @return [type] [description]
 	 */
-	public function view($id)
+	public function view($id_perangkat, $id)
 	{
-		if($id) {
+		if($id && $id_perangkat) {
 			$condition = array('komponen.id'=> $id);
 			$data['model'] = $this->komponen_model->get_one($condition);
 
@@ -156,7 +156,7 @@ class Komponen extends CI_Controller {
 			$this->template->load('main', 'komponen/view', $data);		
 		} else {
 			$this->session->set_flashdata('message_danger', "System Error");
-			redirect('komponen');
+			redirect('perangkat');
 		}
 		
 	}	
@@ -165,20 +165,21 @@ class Komponen extends CI_Controller {
 	 * Action to Delete Record
 	 * @return [type] [description]
 	 */
-	public function delete($id)
+	public function delete($id_perangkat, $id)
 	{
-		if($id) {
+		if($id && $id_perangkat) {
 			$condition = array('id' => $id);
 
 			$res = $this->komponen_model->delete($condition);
 			
 			if($res)
 				$this->session->set_flashdata('message_success', "Successfully deleting data");			
+
+			redirect('perangkat/view/'.$id_perangkat);
+		} else {
+			$this->session->set_flashdata('message_danger', "System Error");	
+			redirect('perangkat');
 		}
-
-		$this->session->set_flashdata('message_danger', "System Error");
-
-		redirect('komponen');
 	}
 
 	/**
